# docs/reference/generated/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchIngestRequest": {
        "additionalProperties": false,
        "description": "A batch of transactions; dryRun validates + masks without persisting anything.\n\nItems are accepted as raw camelCase objects (the `TransactionIngestRequest` shape)\nand validated PER ROW in the handler, so one malformed item becomes a rejection\nrather than failing the whole request \u2014 i.e. the batch is partial-accept.",
        "properties": {
          "dryRun": {
            "default": false,
            "description": "When true, validate + mask but persist nothing.",
            "title": "Dryrun",
            "type": "boolean"
          },
          "transactions": {
            "description": "Transactions to ingest (non-empty; validated per row).",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "minItems": 1,
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "transactions"
        ],
        "title": "BatchIngestRequest",
        "type": "object"
      },
      "BatchIngestResponse": {
        "additionalProperties": false,
        "description": "Outcome of a batch ingest: counts, created rows, and bounded sample errors.",
        "properties": {
          "accepted": {
            "description": "Rows inserted (0 on dryRun).",
            "minimum": 0.0,
            "title": "Accepted",
            "type": "integer"
          },
          "dryRun": {
            "description": "Whether this was a non-persisting validation run.",
            "title": "Dryrun",
            "type": "boolean"
          },
          "duplicates": {
            "description": "Rows skipped as already-existing.",
            "minimum": 0.0,
            "title": "Duplicates",
            "type": "integer"
          },
          "rejected": {
            "description": "Rows rejected as invalid.",
            "minimum": 0.0,
            "title": "Rejected",
            "type": "integer"
          },
          "sampleErrors": {
            "description": "Bounded sample of per-row rejections (PHI-free).",
            "items": {
              "$ref": "#/components/schemas/IngestRejection"
            },
            "title": "Sampleerrors",
            "type": "array"
          },
          "transactions": {
            "description": "Created transactions (empty on dryRun).",
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "accepted",
          "duplicates",
          "rejected",
          "dryRun"
        ],
        "title": "BatchIngestResponse",
        "type": "object"
      },
      "ClientErrorReport": {
        "additionalProperties": false,
        "description": "The frontend client-error sink body (PHI-scrubbed before logging).",
        "properties": {
          "context": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional safe key/value context (scrubbed).",
            "title": "Context"
          },
          "message": {
            "description": "Client error message (scrubbed).",
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ClientErrorReport",
        "type": "object"
      },
      "CsvUploadResponse": {
        "additionalProperties": false,
        "description": "Outcome of a CSV upload: the recorded job id, counts, and bounded sample errors.",
        "properties": {
          "accepted": {
            "description": "Rows inserted.",
            "minimum": 0.0,
            "title": "Accepted",
            "type": "integer"
          },
          "duplicates": {
            "description": "Rows skipped as already-existing.",
            "minimum": 0.0,
            "title": "Duplicates",
            "type": "integer"
          },
          "jobId": {
            "description": "The csv_import job_executions id for this upload.",
            "title": "Jobid",
            "type": "string"
          },
          "rejected": {
            "description": "Rows rejected as invalid.",
            "minimum": 0.0,
            "title": "Rejected",
            "type": "integer"
          },
          "sampleErrors": {
            "description": "Bounded sample of per-row rejections (PHI-free).",
            "items": {
              "$ref": "#/components/schemas/IngestRejection"
            },
            "title": "Sampleerrors",
            "type": "array"
          }
        },
        "required": [
          "jobId",
          "accepted",
          "duplicates",
          "rejected"
        ],
        "title": "CsvUploadResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RiskBand": {
        "description": "Ordinal risk classification assigned to a scored transaction.",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskBand",
        "type": "string"
      },
...
      "TransactionIngestRequest": {
        "additionalProperties": false,
        "description": "One transaction to ingest; structural checks only (semantics in build_canonical).",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "description": "Transaction amount (must be positive).",
            "title": "Amount"
          },
          "channel": {
            "description": "Origination channel.",
            "maxLength": 128,
            "minLength": 1,
            "title": "Channel",
            "type": "string"
          },
          "country": {
            "description": "ISO-3166 alpha-2 country.",
            "maxLength": 2,
            "minLength": 2,
            "title": "Country",
            "type": "string"
          },
          "currency": {
            "description": "ISO-4217 currency code.",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "destAccount": {
            "description": "Destination account (masked).",
            "maxLength": 128,
            "minLength": 1,
            "title": "Destaccount",
            "type": "string"
          },
          "externalId": {
            "description": "Caller-unique id (dedup key).",
            "maxLength": 128,
            "minLength": 1,
            "title": "Externalid",
            "type": "string"
          },
          "features": {
            "additionalProperties": true,
            "description": "Optional extra model features (stored in JSONB).",
            "title": "Features",
            "type": "object"
          },
          "occurredAt": {
            "description": "When the transaction occurred (not future).",
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "originAccount": {
            "description": "Origin account (masked at rest).",
            "maxLength": 128,
            "minLength": 1,
            "title": "Originaccount",
            "type": "string"
          }
        },
        "required": [
          "externalId",
          "amount",
          "currency",
          "occurredAt",
          "originAccount",
          "destAccount",
          "channel",
          "country"
        ],
        "title": "TransactionIngestRequest",
        "type": "object"
      },
      "TransactionListResponse": {
        "additionalProperties": false,
        "description": "A page of transactions plus the opaque cursor for the next page (null at the end).",
        "properties": {
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor for the next page, or null when exhausted.",
            "title": "Nextcursor"
          },
          "transactions": {
            "description": "The transactions on this page (masked).",
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "title": "TransactionListResponse",
        "type": "object"
      },
      "TransactionResponse": {
        "additionalProperties": false,
        "description": "A persisted transaction; account identifiers are masked, riskBand null at ingest.",
        "properties": {
          "agencyId": {
            "description": "Owning tenant (agency) id.",
            "title": "Agencyid",
            "type": "string"
          },
          "amount": {
            "description": "Transaction amount.",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "channel": {
            "description": "Origination channel.",
            "title": "Channel",
            "type": "string"
          },
          "country": {
            "description": "ISO-3166 alpha-2 country code.",
            "title": "Country",
            "type": "string"
          },
          "currency": {
            "description": "ISO-4217 currency code.",
            "title": "Currency",
            "type": "string"
          },
          "destAccount": {
            "description": "Masked destination account identifier.",
            "title": "Destaccount",
            "type": "string"
          },
          "externalId": {
            "description": "Caller-supplied unique id.",
            "title": "Externalid",
            "type": "string"
          },
          "ingestedAt": {
            "description": "When the transaction was ingested.",
            "format": "date-time",
            "title": "Ingestedat",
            "type": "string"
          },
          "latestRunId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Latest analysis run id, if any.",
            "title": "Latestrunid"
          },
          "occurredAt": {
            "description": "When the transaction occurred.",
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "originAccount": {
            "description": "Masked origin account identifier.",
            "title": "Originaccount",
            "type": "string"
          },
          "riskBand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Risk band (null until scored).",
            "title": "Riskband"
          },
          "transactionId": {
            "description": "The transaction's unique id (UUID).",
            "title": "Transactionid",
            "type": "string"
          }
        },
        "required": [
          "transactionId",
          "externalId",
          "agencyId",
          "amount",
          "currency",
          "occurredAt",
          "originAccount",
          "destAccount",
          "channel",
          "country",
          "ingestedAt"
        ],
        "title": "TransactionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "FraudLens AML fraud investigation API (walking skeleton).",
    "title": "FraudLens",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/telemetry/client-error": {
      "post": {
        "description": "Scrub + log a frontend client-error report (PHI-masked, truncated); return 202.",
        "operationId": "report_client_error_api_v1_telemetry_client_error_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientErrorReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Report Client Error",
        "tags": [
          "telemetry"
        ]
      }
    },
    "/api/v1/transactions": {
      "get": {
        "description": "Return a keyset page of the agency's transactions (newest first), optional riskBand.",
        "operationId": "list_transactions_api_v1_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "riskBand",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RiskBand"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Riskband"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Transactions",
        "tags": [
          "transactions"
        ]
      },
      "post": {
        "description": "Ingest one transaction (201); 409 when its externalId already exists for the agency.",
        "operationId": "ingest_transaction_api_v1_transactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Transaction",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/batch": {
      "post": {
        "description": "Ingest a batch (partial-accept); dryRun validates + masks without persisting.",
        "operationId": "ingest_batch_api_v1_transactions_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchIngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Batch",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/upload": {
      "post": {
        "description": "Ingest a text/csv upload (partial-accept); enforce size/row caps; record a job.",
        "operationId": "upload_csv_api_v1_transactions_upload_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsvUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Csv",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{transaction_id}": {
      "get": {
        "description": "Return one transaction by id; 404 when missing or owned by another agency.",
        "operationId": "get_transaction_api_v1_transactions__transaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Transaction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Transaction",
        "tags": [
          "transactions"
        ]
      }
    },
...
  }
}